fixed: Compile warnings (double->float int64->uint truncation)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@35532 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogBoxBase.cpp
@@ -104,6 +104,6 @@ CStdString CGUIDialogBoxBase::GetLocalized(const CVariant &var) const
   if (var.isString())
     return var.asString();
   else if (var.isInteger() && var.asInteger())
-    return g_localizeStrings.Get(var.asInteger());
+    return g_localizeStrings.Get((uint32_t)var.asInteger());
   return "";
 }
